rename to thumbnail-generator
